Corrections to make test/test.rb work
