+++ test/test.rb
@@ -1,6 +1,6 @@
 load '../lib/refine.rb'
 
-prj = Refine.new('date cleanup', 'dates.txt')
+prj = Refine.new({ "project_name" => 'date cleanup', "file_name" => 'dates.csv' })
 prj.apply_operations('operations.json')
 puts prj.export_rows('csv')
 prj.delete_project
